back_end/microserviceGIR/src/main/webapp/app/entities/microserviceGIR/domaine/delete/domaine-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { DomaineService } from '../service/domaine.service';

import { DomaineDeleteDialogComponent } from './domaine-delete-dialog.component';

describe('Domaine Management Delete Component', () => {
  let comp: DomaineDeleteDialogComponent;
  let fixture: ComponentFixture<DomaineDeleteDialogComponent>;
  let service: DomaineService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, DomaineDeleteDialogComponent],
      providers: [NgbActiveModal],
    })
      .overrideTemplate(DomaineDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(DomaineDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(DomaineService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      }),
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
